@@ -31,27 +40,22 @@
     "url": "https://github.com/vuejs/vitepress/issues"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=14.6.0"
   },
   "scripts": {
-    "dev": "run-s dev-shared dev-start",
+    "dev": "rimraf dist && run-s dev-shared dev-start",
     "dev-start": "run-p dev-client dev-node dev-watch",
     "dev-client": "tsc -w -p src/client",
-    "dev-node": "tsc -w -p src/node",
+    "dev-node": "DEV=true pnpm run build-node -w",
     "dev-shared": "node scripts/copyShared",
     "dev-watch": "node scripts/watchAndCopy",
-    "build": "run-s build-prepare build-client build-node build-types",
-    "build-prepare": "rimraf -rf dist && node scripts/copyShared",
+    "build": "run-s build-prepare build-client build-node",
+    "build-prepare": "rimraf dist && node scripts/copyShared",
     "build-client": "tsc -p src/client && node scripts/copyClient",
-    "build-node": "rollup -c scripts/rollup.config.js",
-    "build-types": "run-s build-types-client build-types-node",
-    "build-types-client": "tsc -p src/client --declaration --emitDeclarationOnly --outDir dist/temp && api-extractor run -c api-extractor.client.json && rimraf dist/temp",
-    "build-types-node": "tsc -p src/node --declaration --emitDeclarationOnly --outDir dist/temp && api-extractor run -c api-extractor.node.json && rimraf dist/temp",
-    "lint": "run-s lint:js lint:ts",
-    "lint:js": "prettier --check --write \"{bin,docs,scripts,src}/**/*.js\"",
-    "lint:ts": "prettier --check --write --parser typescript \"{__tests__,src,docs,types}/**/*.ts\"",
-    "test": "vitest run __tests__ -c __tests__/vitest.config.js --global",
-    "test:all": "run-s lint test",
+    "build-node": "rollup --config rollup.config.ts --configPlugin esbuild",
+    "format": "prettier --write .",
+    "lint": "pnpm run format",
+    "test": "vitest -r __tests__",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "release": "node scripts/release.js",
     "docs": "run-p dev docs-dev",
@@ -129,16 +121,20 @@
     "minimist": "^1.2.5",
     "npm-run-all": "^4.1.5",
     "ora": "^5.4.0",
+    "picocolors": "^1.0.0",
     "polka": "^0.5.2",
     "prettier": "^2.3.0",
+    "prompts": "^2.4.2",
     "rimraf": "^3.0.2",
     "rollup": "^2.56.3",
+    "rollup-plugin-dts": "^4.2.2",
     "rollup-plugin-esbuild": "^4.8.2",
     "semver": "^7.3.5",
+    "simple-git-hooks": "^2.7.0",
     "sirv": "^1.0.12",
-    "typescript": "^4.6.4",
-    "vitest": "^0.10.4",
-    "yorkie": "^2.0.0"
+    "supports-color": "^9.2.2",
+    "typescript": "^4.7.2",
+    "vitest": "^0.10.4"
   },
   "pnpm": {
     "peerDependencyRules": {
