@@ -1,22 +1,24 @@
-const fs = require('fs')
-const path = require('path')
-const chalk = require('chalk')
-const semver = require('semver')
-const { prompt } = require('enquirer')
-const execa = require('execa')
-const currentVersion = require('../package.json').version
+import { readFileSync, writeFileSync } from 'fs'
+import { resolve } from 'path'
+import c from 'picocolors'
+import { inc as _inc, valid } from 'semver'
+import prompts from 'prompts'
+import { execa } from 'execa'
+import { version as currentVersion } from '../package.json'
+import { fileURLToPath } from 'url'
 
 const versionIncrements = ['patch', 'minor', 'major']
 
-const inc = (i) => semver.inc(currentVersion, i)
+const dir = dirname(fileURLToPath(import.meta.url))
+const inc = (i) => _inc(currentVersion, i)
 const run = (bin, args, opts = {}) =>
   execa(bin, args, { stdio: 'inherit', ...opts })
-const step = (msg) => console.log(chalk.cyan(msg))
+const step = (msg) => console.log(c.cyan(msg))
 
 async function main() {
   let targetVersion
 
-  const { release } = await prompt({
+  const { release } = await prompts({
     type: 'select',
     name: 'release',
     message: 'Select release type',
@@ -25,7 +27,7 @@ async function main() {
 
   if (release === 'custom') {
     targetVersion = (
-      await prompt({
+      await prompts({
         type: 'input',
         name: 'version',
         message: 'Input custom version',
@@ -36,11 +38,11 @@ async function main() {
     targetVersion = release.match(/\((.*)\)/)[1]
   }
 
-  if (!semver.valid(targetVersion)) {
+  if (!valid(targetVersion)) {
     throw new Error(`Invalid target version: ${targetVersion}`)
   }
 
-  const { yes: tagOk } = await prompt({
+  const { yes: tagOk } = await prompts({
     type: 'confirm',
     name: 'yes',
     message: `Releasing v${targetVersion}. Confirm?`
@@ -63,7 +65,7 @@ async function main() {
   await run('pnpm', ['changelog'])
   await run('pnpm', ['prettier', '--write', 'CHANGELOG.md'])
 
-  const { yes: changelogOk } = await prompt({
+  const { yes: changelogOk } = await prompts({
     type: 'confirm',
     name: 'yes',
     message: `Changelog generated. Does it look good?`
@@ -90,12 +92,12 @@ async function main() {
 }
 
 function updatePackage(version) {
-  const pkgPath = path.resolve(path.resolve(__dirname, '..'), 'package.json')
-  const pkg = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'))
+  const pkgPath = resolve(resolve(dir, '..'), 'package.json')
+  const pkg = JSON.parse(readFileSync(pkgPath, 'utf-8'))
 
   pkg.version = version
 
-  fs.writeFileSync(pkgPath, JSON.stringify(pkg, null, 2) + '\n')
+  writeFileSync(pkgPath, JSON.stringify(pkg, null, 2) + '\n')
 }
 
 main().catch((err) => console.error(err))
