@@ -1,10 +1,11 @@
-import chalk from 'chalk'
+import c from 'picocolors'
 import minimist from 'minimist'
 import { createServer, build, serve } from '.'
+import { version } from '../../package.json'
 
 const argv: any = minimist(process.argv.slice(2))
 
-console.log(chalk.cyan(`vitepress v${require('../../package.json').version}`))
+console.log(c.cyan(`vitepress v${version}`))
 
 const command = argv._[0]
 const root = argv._[command ? 1 : 0]
@@ -20,20 +21,20 @@ if (!command || command === 'dev') {
       server.printUrls()
     })
     .catch((err) => {
-      console.error(chalk.red(`failed to start server. error:\n`), err)
+      console.error(c.red(`failed to start server. error:\n`), err)
       process.exit(1)
     })
 } else if (command === 'build') {
   build(root, argv).catch((err) => {
-    console.error(chalk.red(`build error:\n`), err)
+    console.error(c.red(`build error:\n`), err)
     process.exit(1)
   })
 } else if (command === 'serve') {
   serve(argv).catch((err) => {
-    console.error(chalk.red(`failed to start server. error:\n`), err)
+    console.error(c.red(`failed to start server. error:\n`), err)
     process.exit(1)
   })
 } else {
-  console.log(chalk.red(`unknown command "${command}".`))
+  console.log(c.red(`unknown command "${command}".`))
   process.exit(1)
 }
