@@ -1,7 +1,7 @@
 import path from 'path'
 import fs from 'fs-extra'
-import chalk from 'chalk'
-import globby from 'globby'
+import c from 'picocolors'
+import fg from 'fast-glob'
 import {
   normalizePath,
   AliasOptions,
@@ -23,7 +23,6 @@ import { MarkdownOptions } from './markdown/markdown'
 import _debug from 'debug'
 
 export { resolveSiteDataByRoute } from './shared'
-export type { MarkdownOptions }
 
 const debug = _debug('vitepress:config')
 
@@ -125,14 +124,14 @@ export async function resolveConfig(
     ? userThemeDir
     : DEFAULT_THEME_PATH
 
-  // Important: globby/fast-glob doesn't guarantee order of the returned files.
+  // Important: fast-glob doesn't guarantee order of the returned files.
   // We must sort the pages so the input list to rollup is stable across
   // builds - otherwise different input order could result in different exports
   // order in shared chunks which in turns invalidates the hash of every chunk!
   // JavaScript built-in sort() is mandated to be stable as of ES2019 and
   // supported in Node 12+, which is required by Vite.
   const pages = (
-    await globby(['**.md'], {
+    await fg(['**.md'], {
       cwd: srcDir,
       ignore: ['**/node_modules', ...(userConfig.srcExclude || [])]
     })
@@ -190,7 +189,7 @@ async function resolveUserConfig(
     : {}
 
   if (configPath) {
-    debug(`loaded config at ${chalk.yellow(configPath)}`)
+    debug(`loaded config at ${c.yellow(configPath)}`)
   } else {
     debug(`no config file found.`)
   }
